+++ src/consumer/bit-map/component-map.js
@@ -2,7 +2,7 @@
 import path from 'path';
 import logger from '../../logger/logger';
 import { COMPONENT_ORIGINS, BIT_MAP } from '../../constants';
-import { pathNormalizeToLinux, pathJoinLinux, pathRelativeLinux } from '../../utils';
+import { pathNormalizeToLinux, pathJoinLinux, pathRelativeLinux, isValidPath } from '../../utils';
 import type { PathLinux, PathOsBased } from '../../utils/path';
 
 export type ComponentOrigin = $Keys<typeof COMPONENT_ORIGINS>;
@@ -171,13 +171,9 @@ export default class ComponentMap {
   }
 
   validate() {
-    const isPathValid = (pathStr) => {
-      if (pathStr.startsWith('./') || pathStr.startsWith('../') || pathStr.includes('\\')) return false;
-      return true;
-    };
     const errorMessage = `failed adding a component-map record (to ${BIT_MAP} file).`;
     if (!this.mainFile) throw new Error(`${errorMessage} mainFile attribute is missing`);
-    if (!isPathValid(this.mainFile)) {
+    if (!isValidPath(this.mainFile)) {
       throw new Error(`${errorMessage} mainFile attribute ${this.mainFile} is invalid`);
     }
     // if it's an environment component (such as compiler) the rootDir is an empty string
@@ -185,7 +181,7 @@ export default class ComponentMap {
       throw new Error(`${errorMessage} rootDir attribute is missing`);
     }
     // $FlowFixMe
-    if (this.rootDir && !isPathValid(this.rootDir)) {
+    if (this.rootDir && !isValidPath(this.rootDir)) {
       throw new Error(`${errorMessage} rootDir attribute ${this.rootDir} is invalid`);
     }
     if (this.rootDir && this.origin === COMPONENT_ORIGINS.AUTHORED) {
@@ -196,7 +192,7 @@ export default class ComponentMap {
     }
     if (!this.files || !this.files.length) throw new Error(`${errorMessage} files list is missing`);
     this.files.forEach((file) => {
-      if (!isPathValid(file.relativePath)) {
+      if (!isValidPath(file.relativePath)) {
         throw new Error(`${errorMessage} file path ${file.relativePath} is invalid`);
       }
     });